services/ai_service/pyproject.toml
[tool.poetry]
name = "ai-service"
version = "0.1.0"
description = "AI Service for Tasks & Finance Tracker"
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
faststream = {extras = ["rabbit"], version = "^0.5.0"}
langchain = "^0.1.20"
langchain-google-genai = "^0.0.5"
langchain-openai = "^0.0.5"
langchain-groq = "^0.1.3"
langchain-community = "^0.0.38"
pyjwt = "^2.8.0"
pydantic-settings = "^2.1.0"
pydantic = "^2.5.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
httpx = "^0.27.0"
tavily-python = "^0.3.0"
redis = "^5.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
